requirements.txt
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==3.0.2
PyYAML==6.0.2
Jinja2==3.1.6
Markdown==3.7
marko==2.1.2
MarkupSafe==3.0.2
PyYAML==6.0.2
